Move remaining test utilities into xcp_d._testing
Move download_test_data, get_test_data_path, check_generated_files,
check_affines, run_command, reorder_expected_outputs, and list_files
from test/utils.py into src/xcp_d/_testing.py. Delete test/utils.py.

Update get_test_data_path to compute test/data/ relative to the new
location of _testing.py (src/xcp_d/ -> up 2 levels -> test/data/).
