+++ test/test_workflows_anatomical.py
@@ -5,9 +5,8 @@ import shutil
 
 import pytest
 
-from test.utils import get_nodes
 from xcp_d import config
-from xcp_d._testing import mock_config
+from xcp_d._testing import get_nodes, mock_config
 from xcp_d.workflows import anatomical
 from xcp_d.workflows.base import clean_datasinks
